Create metaphone_step2.py
This code is function, but disjointed. I have to figure out how it plugs into the main script. Basically, two names need to be plugged into the compare function. to do this, I may elect to set a form of indexing to associate the closest phonetically spelled names to compare. This would take an incredibly long time on large datasets.